requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
filelock==3.18.0
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
PyAudio==0.2.14
regex==2025.7.31
requests==2.32.4
setuptools==80.9.0
SpeechRecognition==3.14.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
